[#5011] Add catalog utility libraries
Adds a new catalog library for general catalog transaction execution.

Adds a new catalog utilities library for adding more specific operations.
